refactor: big refactor for commander module

+++ internal/service/sensor/scrapper/scrapper.go
@@ -66,7 +66,7 @@ func Scrap() []Sensor {
 	return sensors
 }
 
-// TODO Maybe not a place for it
+// FilterSensorsByHourDiff TODO Maybe not a place for it
 func FilterSensorsByHourDiff(sensors []Sensor, diffInHours int) []Sensor {
 	var aliveSensors []Sensor
 	layout := "2006-01-02T15:04:05.999999999Z"
@@ -78,7 +78,7 @@ func FilterSensorsByHourDiff(sensors []Sensor, diffInHours int) []Sensor {
 		}
 
 		diffInHours := sensorTime.Sub(time.Now().UTC()).Hours()
-		if diffInHours > float64(-1*diffInHours) {
+		if diffInHours > -1*diffInHours {
 			aliveSensors = append(aliveSensors, sensor)
 		}
 	}
